% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{FuzzyCluster-class}
\alias{FuzzyCluster-class}
\title{An S4 class to represent a Cluster}
\description{
An S4 class to represent a Cluster
}
\section{Slots}{

\describe{
\item{\code{data}}{a \code{data.frame} with numerical data. Each column corresponds to a dimension of a \code{n}-dimensional clustering domain.}

\item{\code{U}}{tmp}

\item{\code{W}}{tmp}

\item{\code{centroids}}{tmp}

\item{\code{phi}}{a numerical value representing the degree of fuzziness or overlap of the generated clusters.}

\item{\code{classes}}{numerical (integer) vector with the number of cluster to generate.}

\item{\code{distance}}{used to calculate the clustering. Possible values are: 'Euclidean', 'Diagonal', and 'Mahalanobis'.}

\item{\code{alpha}}{tmp}

\item{\code{Ue_mean-Ue_req}}{tmp}

\item{\code{iterations}}{total number of iterations}

\item{\code{pred_int}}{tmp}
}}

