% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuzzyCluster_methods.R
\docType{methods}
\name{FuzzyCluster-methods}
\alias{FuzzyCluster-methods}
\title{Methods for FuzzyCluster class}
\usage{
\S4method{plot}{FuzzyCluster}(x, .plot = NULL, ...)

\S4method{summary}{FuzzyCluster}(object, extragrade_label = "Eg",
  decimals = 2, latex = F, caption = NULL, label = NULL)

\S4method{vcov}{FuzzyCluster}(object)

\S4method{predict}{FuzzyCluster}(object, newdata)
}
\arguments{
\item{x, object}{FuzzyCluster object}

\item{.plot}{numeric (optional). If just one plot is required it can be specified here.}

\item{...}{arguments to be passed to subsequent methods, e.g., plot.FuzzyCluster for the plot method}

\item{extragrade_label}{string to use as label for the extragrade class. Default set to \code{'Eg' }.}

\item{decimals}{number of decimals to display.}

\item{latex}{if \code{FALSE} (default) the summary is displayed as a regular R output. If \code{TRUE}, the output is formated as a latex table (using booktabs).}

\item{caption}{if \code{latex} is set to \code{TRUE}, is the sting used as a caption for the table.}

\item{label}{if \code{latex} is set to \code{TRUE}, is the sting used as a label for the table.}

\item{newdata}{a \code{data.frame} object in which to look for variables with which to predict.}
}
\description{
Common methods implemented for the FuzzyCluster class.
}
\examples{
### Create FuzzyClusterGroup
fkm <- fuzzyRun(cars,1.5,2:4,3)

### Functions to be used on part of the FuzzyClusterGroup
plot(fkm@clusters$`2`,.plot=1)

summary(fkm@clusters$`2`,latex=TRUE)
}

