% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{FuzzyClusterGroup-class}
\alias{FuzzyClusterGroup-class}
\title{An S4 class to represent a group of FuzzyCluster objects}
\description{
An S4 class to represent a group of FuzzyCluster objects
}
\section{Slots}{

\describe{
\item{\code{clusters}}{A list containing FuzzyCluster objects}

\item{\code{confidence}}{Confidence level for all the FuzzyCluster objects}
}}

