% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fuzzyRun.R
\name{fuzzyRun}
\alias{fuzzyRun}
\title{Fuzzy k-means with extragrade sequence}
\usage{
fuzzyRun(data, phi, classes, disttype = 3, exp_eg = NULL, obs = NULL,
  pred = NULL, .conf = 0.95, alpha_values = NULL)
}
\arguments{
\item{data}{a \code{data.frame} with numerical data. Each column corresponds to a dimension of a \code{n}-dimensional clustering domain.}

\item{phi}{a numerical value representing the degree of fuzziness or overlap of the generated clusters.}

\item{classes}{numerical (integer) vector with the number of cluster to generate.}

\item{disttype}{numerical value representing the type of distance to calculate for the clustering. Possible values are: 1 (Euclidean), 2 (Diagonal), and 3 (default, Mahalanobis).}

\item{exp_eg}{the expected fraction of extragrades. If not provided, it is assumed to be dependant on the number of clusters (\code{1/(1+nclusters)}).}

\item{obs}{numerical vector (optional) corresponding to the observed values of a modeled variable. It is used in conjunction with the \code{pred} argument to estimimate the \code{\link{prediction_interval}} for each \code{classes}.}

\item{pred}{numerical vector (optional) corresponding to the predicted values of a modeled variable. It is used in conjunction with the \code{obs} argument to estimimate the \code{\link{prediction_interval}} for each \code{classes}.}

\item{.conf}{confidence level used when estimimating the \code{\link{prediction_interval}}. Default value: 0.95.}

\item{alpha_values}{numerical vector (optional), corresponding to optimised alpha values. This function estimates optimum alpha values, so this argumet is mostly used at developing.}
}
\value{
\code{FuzzyClusterGroup}
}
\description{
This function generates an FuzzyClusterGroup for a sequence of cluster numbers.
}
\examples{
fkm <- fuzzyRun(cars,1.5,2:6,3)
}
\references{
De Gruijter, J. and McBratney, A. 1988. A modified fuzzy k-means method for predictive classification. In: 1st Conference of the International Federation of Classification Societies. Ed. by H. Bock. Elsevier Science, Amsterdam: pp. 97–104.
}

