% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mahaldist}
\alias{mahaldist}
\title{Calculate distances}
\usage{
mahaldist(x, y, w)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{w}{matrix}
}
\value{
distance matrix
}
\description{
Calculate distances
}

