% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prediction_interval.R
\name{prediction_interval}
\alias{prediction_interval}
\title{Prediction interval estimation}
\usage{
prediction_interval(clust_data, obs = NULL, pred = NULL, .conf = 0.95)
}
\arguments{
\item{clust_data}{a \code{FuzzyCluster} object.}

\item{obs}{numerical vector (optional) corresponding to the observed values of a modeled variable. It is used in conjunction with the \code{pred} argument to estimimate the prediction interval for each \code{classes}.}

\item{pred}{numerical vector (optional) corresponding to the predicted values of a modeled variable. It is used in conjunction with the \code{obs} argument to estimimate the prediction interval for each \code{classes}.}

\item{.conf}{confidence level used when estimimating the prediction interval. Default value: 0.95.}
}
\value{
\code{list}
}
\description{
This function estimates prediction intervals for a FuzzyCluster object. It shouldn't be neccesary to use it by itself because it is internally used by \code{\link{fuzzyRun}}.
}

