% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{data}
\name{soilsecuritylab.env}
\alias{soilsecuritylab.env}
\title{Create a new environment for the package}
\format{An object of class \code{environment} of length 0.}
\usage{
soilsecuritylab.env
}
\description{
Create a new environment for the package
}
\keyword{datasets}

