% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FuzzyClusterGroup_methods.R
\docType{methods}
\name{FuzzyClusterGroup-methods}
\alias{FuzzyClusterGroup-methods}
\title{Methods for FuzzyClusterGroup class}
\usage{
\S4method{plot}{FuzzyClusterGroup}(x, ...)

\S4method{length}{FuzzyClusterGroup}(x)
}
\arguments{
\item{x}{FuzzyClusterGroup object}

\item{...}{arguments to be passed to subsequent methods, e.g., plot.FuzzyCluster for the plot method}
}
\description{
Common methods implemented for the FuzzyClusterGroup class.
}

